Clean up some sign/size issues.

+++ src/net.c
@@ -46,8 +46,8 @@ int server_connect(const char *host, const char *port) {
 }
 
 
-int send_bytes(const void *buf, const size_t len) {
-    int bytes_sent = send(server_sock, buf, len, 0);
+ssize_t send_bytes(const void *buf, const size_t len) {
+    ssize_t bytes_sent = send(server_sock, buf, len, 0);
     if (bytes_sent == -1)
         die("send() error: %s", strerror(errno));
     return bytes_sent;
